PST Maintenance
Version:  1.2 / Updated:  3/24/2015 / By:  ZT
Change Log:
v0.7 = Made further adjustments and improvements in my scripting.  ZT
v1.0 = During testing, found the Mitel UCA has a process that is opening
the PST files and causing the script to not be able to move them.  Added
in a check for the process and then to stop it.
v1.1 = Updated a couple checks to parse through the script quicker if
there was noting found in the check.  This was also done to clean up log
file entries.  Also fixed an issue where a " quote was left off a few
lines that were added in the last revision.
v1.2 = Loop was created in script, forgot the counting variable in the
new While loop I added.  Modified output so that it writes to screen and
file to make it easier to see on end user machines what's going on.